Send 403

+++ delete.php
@@ -12,7 +12,7 @@ if (!isset($_SESSION["user"])){
 
 $id = $_GET["id"];
 
-$statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id");
+$statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id LIMIT 1");
 $statement->execute([":id" => $id]);
 
 if($statement->rowCount() == 0){
@@ -20,6 +20,15 @@ if($statement->rowCount() == 0){
     echo("HTTP 404 NOT FOUND");
     return;
 }
+
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]){
+  http_response_code(403);
+  echo("HTTP 403 UNAUTHORIZED");
+  return;
+}
+
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
 
 header("Location: home.php");
